+++ src/main/react/src/components/filingApp/TaxReturnStatus.tsx
@@ -1,0 +1,19 @@
+import { CardGroup, Card, CardHeader, CardBody } from "@trussworks/react-uswds"
+
+export default function TaxReturnStatus() {
+    return <>
+        <CardGroup>
+            <Card gridLayout={{ col: 12 }}>
+                <CardHeader>
+                    <h3 className="usa-card__heading">Tax calculations</h3>
+                </CardHeader>
+                <CardBody>
+                    <div>
+                        "asdf"
+                    </div>
+                </CardBody>
+            </Card>
+        </CardGroup>
+    </>
+}
+
